BCOP-1643 - Adjust impressions for snapshot in inventories
git-tfs-id: [http://devvm01tfs03.telamericamedia.com:8080/tfs/CMWPA]$/Telamerica/Broadcast/OnDeck;C20744

+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/OpenMarketInventoryAllocationSnapshotDto.cs
@@ -7,7 +7,7 @@
         public int Spots { get; set; }
         public string CreatedBy { get; set; }
         public string Isci { get; set; }
-        public int Impressions { get; set; }
+        public double Impressions { get; set; }
         public decimal SpotCost { get; set; }
         public int SpotLengthId { get; set; }
         public short StationCode { get; set; }
